Move kvs_client.hpp include file to client
