+++ CMakeLists.txt
@@ -63,6 +63,7 @@ INCLUDE_DIRECTORIES(${SPDLOG_INCLUDE_DIRS})
 INCLUDE_DIRECTORIES(${ZEROMQCPP_INCLUDE_DIRS})
 INCLUDE_DIRECTORIES(${ZEROMQ_INCLUDE_DIRS})
 INCLUDE_DIRECTORIES(${YAMLCPP_INCLUDE_DIRS})
+INCLUDE_DIRECTORIES(client/cpp)
 INCLUDE_DIRECTORIES(common/include)
 INCLUDE_DIRECTORIES(include)
