+++ client/cpp/cli.cpp
@@ -14,7 +14,7 @@
 
 #include <fstream>
 
-#include "client/kvs_client.hpp"
+#include "kvs_client.hpp"
 #include "yaml-cpp/yaml.h"
 
 #include <assert.h>
