+++ client/cpp/kvs_client.hpp
